♻️ fetchAccountProviders関数を切り出し
