+++ src/app/_components/AccountLinkingSection/AccountLinkingSection.tsx
@@ -1,25 +1,11 @@
 import { headers } from 'next/headers';
 import { AccountLinkingToggle } from '@/app/_components/AccountLinkingSection/AccountLinkingToggle';
-
-export type AccountProvider = 'google' | 'line';
-
-type ProviderObject = {
-  provider: AccountProvider;
-};
+import {
+  AccountProvider,
+  fetchAccountProviders,
+} from '@/lib/fetchAccountProviders';
 
 const Accounts: AccountProvider[] = ['google', 'line'];
-
-async function fetchAccountProviders(
-  cookie: string
-): Promise<ProviderObject[]> {
-  const res = await fetch(`${process.env.NEXTAUTH_URL}/api/account`, {
-    method: 'GET',
-    headers: { cookie },
-    next: { tags: ['account'] },
-  });
-  if (!res.ok) return [];
-  return res.json();
-}
 
 const checkSwitchDisabled = (
   provider: AccountProvider,
